Move persona components to src/components/persona
自己紹介(Persona) 関連のコンポーネントを src/components/Persona ディレクトリに移動しました
