+++ src/components/persona/PersonaPrevCard.module.css
@@ -15,6 +15,8 @@
 }
 
 .DisplayName {
+    padding-bottom: .2rem;
+    border-bottom: 1px #666 solid;
     margin-bottom: 0.5rem;
 }
