+++ tests/components/PersonaTestPage.tsx
@@ -1,6 +1,6 @@
-import PersonaPrevCard from "@/components/PersonaPrevCard";
-import PersonaPrevBanner from "@/components/PersonaPrevBanner";
-import PersonaFullPage from "@/components/PersonaFullPage";
+import PersonaPrevCard from "@/components/persona/PersonaPrevCard";
+import PersonaPrevBanner from "@/components/persona/PersonaPrevBanner";
+import PersonaFullPage from "@/components/persona/PersonaFullPage";
 import {Persona} from "@/types/Persona";
 import {TestPersona} from "./testPersona";
 import styles from "./PersonaTestPage.module.css"
